Add --talk-name=org.freedesktop.Flatpak
This addresses #14 by making the org.freedesktop.Flatpak name
available to Emacs. This weakens the sandbox by making the
flatpak-spawn --host command work from within Emacs. This is expected
to not be harmful, because Emacs would typically have full access to
practically all the files of the user in any case.